requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
django-admin-thumbnails==0.2.8
django-axes==6.3.0
idna==3.6
Pillow==10.1.0
psycopg==3.1.16
python-decouple==3.8
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.1.0
